refactor: move helper to domain/course
